fixed about me
"I learned early on in my childhood the joys of how art and beauty bring to people. I constantly sought out new techniques to experiment with art. In the process of discovery, I grew through the struggles of failure and adaptation. College led to graphic design, where I learned the joy of storytelling, communication, and storytelling. Today, I continuously strive to create work that breaks away from fleeting trends to build a lasting design. I believe that effective communication is the most important aspect of any brand. This mission is united with my love for studying the language and drive of brands. In my free time, I visit art museums and daydream about Porsche 911 and Taycan. "